+++ src/scripts/middlewares/logger.js
@@ -1,0 +1,12 @@
+const logger = store => next => action => {
+  console.group(action.type)
+  console.info('dispatching', action)
+  const result = next(action)
+
+  console.log('next state', store.getState())
+  console.groupEnd(action.type)
+
+  return result
+}
+
+export default logger
